# openapi.yaml
openapi: 3.0.3
info:
  title: Task Management API
  description: API documentation for the Task Management web application, featuring task tracking, team collaboration, and real-time messaging.
  version: 1.0.0
servers:
  - url: /api/v1
    description: Base API path
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: refresh_token
  schemas:
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    LoginResponse:
      type: object
      properties:
        user_id:
          type: integer
        email:
          type: string
        user_name:
          type: string
        access_token:
          type: string
        refresh_token:
          type: string
          nullable: true
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - full_name
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        full_name:
          type: string
    RegisterResponse:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        full_name:
          type: string
    VerifyOTPRequest:
      type: object
      required:
        - email
        - otp
      properties:
        email:
          type: string
        otp:
          type: string
    VerifyOtpResponse:
      type: object
      properties:
        reset_token:
          type: string
    ResetPasswordRequest:
      type: object
      required:
        - email
        - reset_token
        - new_password
      properties:
        email:
          type: string
        reset_token:
          type: string
        new_password:
          type: string
...
    TaskStatus:
      type: string
      enum: [todo, in_progress, review, done, over_due]
...
    UpdatedTaskRequest:
      type: object
      properties:
        new_due_date:
          type: string
          format: date
        new_status:
          $ref: '#/components/schemas/TaskStatus'
        version:
          type: integer
...
    CreatedCommentRequest:
      type: object
      required:
        - task_id
        - content
      properties:
        task_id:
          type: integer
        content:
          type: string
...
    SendMessageRequest:
      type: object
      required:
        - conversation_id
      properties:
        conversation_id:
          type: integer
        content:
          type: string
        type:
          type: string
          enum: [TEXT, IMAGE, FILE]
        file:
          type: string
          format: binary
...
    DocumentResponse:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
          format: byte
        created_at:
          type: string
          format: date-time
    UpdateDocumentRequest:
      type: object
      properties:
        content:
          type: string
          format: byte
...
paths:
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: User login
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          headers:
            Set-Cookie:
              schema:
                type: string
                example: refresh_token=abc...; HttpOnly; Secure; Path=/api/v1/auth
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: User registration
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
  /api/v1/auth/google:
    post:
      tags: [Auth]
      summary: Google OAuth2 login
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
  /api/v1/auth/facebook:
    post:
      tags: [Auth]
      summary: Facebook OAuth2 Login
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      security:
        - CookieAuth: []
      responses:
        '201':
          description: Token refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: Logout user
      security:
        - CookieAuth: []
      responses:
        '200':
          description: Logout successful
  /api/v1/auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request password reset OTP
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
      responses:
        '200':
          description: OTP sent
  /api/v1/auth/verify-otp:
    post:
      tags: [Auth]
      summary: Verify OTP
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyOTPRequest'
      responses:
        '200':
          description: OTP verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyOtpResponse'
  /api/v1/auth/reset-password:
    post:
      tags: [Auth]
      summary: Reset password
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password reset successful
...
  /api/v1/task/{taskId}:
    patch:
      tags: [Tasks]
      summary: Update task status or due date
      security:
        - BearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatedTaskRequest'
      responses:
        '200':
          description: Task updated
...
  /api/v1/comment:
    post:
      tags: [Comments]
      summary: Create comment
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatedCommentRequest'
      responses:
        '201':
          description: Comment created
    delete:
      tags: [Comments]
      summary: Delete comment
      security:
        - BearerAuth: []
      parameters:
        - name: commentId
          in: query
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Comment deleted
  /api/v1/comment/task/{taskId}:
    get:
      tags: [Comments]
      summary: Get comments for a task
      security:
        - BearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of comments
...
  /api/v1/message:
    post:
      tags: [Messaging]
      summary: Send a message
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '201':
          description: Message sent
  /api/v1/message/conversation/{conversationId}:
    get:
      tags: [Messaging]
      summary: Get messages in a conversation
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
        - name: cursorId
          in: query
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: List of messages
...
  /api/v1/document:
    get:
      tags: [Documents]
      summary: Get document by task ID
      security:
        - BearerAuth: []
      parameters:
        - name: taskId
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Document content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
  /api/v1/document/{documentId}:
    patch:
      tags: [Documents]
      summary: Update document content
      security:
        - BearerAuth: []
      parameters:
        - name: documentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDocumentRequest'
      responses:
        '200':
          description: Document updated